+++ resharper/resharper-unity/test/src/CSharp/Intentions/QuickFixes/RedundantInitializeOnLoadAttributeQuickFixTests.cs
@@ -13,15 +13,19 @@ namespace JetBrains.ReSharper.Plugins.Unity.Tests.CSharp.Intentions.QuickFixes
 
         [Test] public void Test01() { DoNamedTest(); }
         [Test] public void Test02() { DoNamedTest(); }
+
+        // Test06 so we can share files between availability and action tests
+        [Test] public void Test06() { DoNamedTest(); }
     }
 
     [TestUnity]
-    public class RedundantInitializeOnLoadAttributeQuickFixRemoveTests : CSharpQuickFixTestBase<RedundantAttributeDeadCodeQuickFix>
+    public class RedundantInitializeOnLoadAttributeQuickFixRemoveTests : CSharpQuickFixTestBase<RemoveRedundantAttributeQuickFix>
     {
         protected override string RelativeTestDataPath=> @"CSharp\Intentions\QuickFixes\RedundantInitializeOnLoadAttribute";
 
         [Test] public void Test01() { DoNamedTest(); }
         [Test] public void Test02() { DoNamedTest(); }
+        [Test] public void Test06() { DoNamedTest(); }
     }
 
     [TestUnity]