+++ resharper/resharper-unity/test/src/CSharp/Intentions/QuickFixes/RedundantSerializeFieldAttributeQuickFixTests.cs
@@ -14,10 +14,11 @@ namespace JetBrains.ReSharper.Plugins.Unity.Tests.CSharp.Intentions.QuickFixes
         [Test] public void Test03() { DoNamedTest(); }
         [Test] public void Test04() { DoNamedTest(); }
         [Test] public void Test05() { DoNamedTest(); }
+        [Test] public void Test06() { DoNamedTest(); }
     }
 
     [TestUnity]
-    public class RedundantSerializeFieldAttributeQuickFixTests : CSharpQuickFixTestBase<RedundantAttributeDeadCodeQuickFix>
+    public class RedundantSerializeFieldAttributeQuickFixTests : CSharpQuickFixTestBase<RemoveRedundantAttributeQuickFix>
     {
         protected override string RelativeTestDataPath=> @"CSharp\Intentions\QuickFixes\RedundantSerializeFieldAttribute";
 
@@ -26,5 +27,6 @@ namespace JetBrains.ReSharper.Plugins.Unity.Tests.CSharp.Intentions.QuickFixes
         [Test] public void Test03() { DoNamedTest(); }
         [Test] public void Test04() { DoNamedTest(); }
         [Test] public void Test05() { DoNamedTest(); }
+        [Test] public void Test06() { DoNamedTest(); }
     }
 }